ported a problem from adevent of code 2020, added tests and generic mapping for the solve endpoint

+++ src/BE/ProblemSolver/ProblemSolver.Application/CQRS/Euler/GetPrimeSum/GetPrimeSumResult.cs
@@ -2,6 +2,6 @@
 {
     public class GetPrimeSumResult
     {
-        public int Result { get; set; }
+        public long Result { get; set; }
     }
 }
